Extract ExpandableCell into a class
